+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - English
   - French
   - Spanish
+  - Doc Workbench: Add language selector
 - Enhancements
   - Data Workbench: Make more explicit the use of filtering using master queries results (new badge)
   - Org Manager: Add more contextual actions: Freeze / Unfreeze users, Before / After refresh sandbox...